@@ -3,16 +3,16 @@ apply plugin: 'com.novoda.bintray-release'
 
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 101
         versionName "${version}"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -33,9 +33,10 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:support-compat:27.0.2'
-    implementation 'org.slf4j:slf4j-android:1.7.25'
-    implementation 'io.jsonwebtoken:jjwt:0.9.0'
+    implementation 'androidx.core:core:1.1.0'
+    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'org.slf4j:slf4j-android:1.7.30'
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
 
     api 'com.pax.market:paxstore-3rd-app-java-sdk:7.0.0'
 }
